Add editing and interaction ability to tasks
User can now click on a task in PB to view it, and then click on the edit button to edit.
Alternatively, they can click on the edit button in the PB which will automatically start editing.

The user can then save their changes by clicking the "Save" button (which currently only prints the output to console, doesn't actually save yet.)
Drop-down options have also been configured and story points are forced within the range of 1 to 9.